feat(17-02): preserve v9 transformer wire tokens
- Map transformer failure modes to strict and soft_fail before config serialization
- Marshal copied representation metadata with legacy failure-mode tokens
- Keep the version constant at v9 for the static compatibility gate

+++ gin.go
@@ -28,7 +28,7 @@ const (
 	//       iteration of the adaptive string index section before the wire
 	//       format was finalised in v6.
 	//   v4: earlier pre-OSS format
-	Version = uint16(9)
+	Version = 9
 )
 
 const (
